pyproject.toml
[project]
name = "stockland-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "langchain>=1.2.15",
    "langgraph>=1.1.6",
    "loguru>=0.7.3",
    "openai>=2.31.0",
    "pgvector>=0.4.2",
    "playwright>=1.58.0",
    "prefect>=3.6.26",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.12.5",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "reportlab>=4.4.10",
    "requests>=2.33.1",
    "sentence-transformers>=5.4.0",
    "sqlalchemy>=2.0.49",
    "streamlit>=1.56.0",
    "uvicorn>=0.44.0",
]
